Remove SSE and replace with a polling
Now using a background polling approach with refresh only when needed.
