+++ TODO.md
@@ -4,5 +4,4 @@ TODO
 * Make mobile a thing :)
 * Make docker a thing
 * Deal with configuration somehow
-* Websockets auto-refresh (replace SSE)
 * Verify that password resets work in prod
