+++ app/views/dash/_beer_list.html.erb
