+++ app/views/dash/index.html.erb
