--- app/views/streams/dash.js.erb
@@ -1,1 +1,0 @@
-$('#beer-list').html("<%= j(render partial:'beer_list') %>");
