+++ config/routes.rb
@@ -5,19 +5,16 @@ Clubfare::Application.routes.draw do
 	resources :beers
 	resources :brewers
 	resources :users
-	resources :streams
+	resources :dash
 
 	namespace :api do
 		resources :beers
 	end
 
-	root 'streams#dash'
+	root 'dash#index'
 
-	match '/dash',		to: 'streams#dash',			via: 'get'
+	match '/dash',		to: 'dash#index',			via: 'get'
 
-	# Allow SSE streaming
-	get 'streaming' => 'streams#update_stream', as: 'streaming'
-	
 	# Beer label generation
 	match 'beers/:id/label' => 'beers#label', as: :beers_label, via: 'get'
