--- lib/reloader/sse.rb
@@ -1,20 +1,0 @@
-require 'json'
-
-module Reloader
-	class SSE
-		def initialize io
-			@io = io
-		end
-
-		def write object, options = {}
-			options.each do |k,v|
-				@io.write "#{k}: #{v}\n"
-			end
-			@io.write "data: #{JSON.dump(object)}\n\n"
-		end
-
-		def close
-			@io.close
-		end
-	end
-end
